+++ include/libpldm/utils.h
@@ -6,6 +6,7 @@
 extern "C" {
 #endif
 
+#include <libpldm/edac.h>
 #include <libpldm/pldm_types.h>
 
 #include <stdbool.h>
@@ -21,31 +22,6 @@ struct variable_field {
 	const uint8_t *ptr;
 	size_t length;
 };
-
-/** @brief Compute Crc8(same as the one used by SMBUS)
- *
- *  @param[in] data - Pointer to the target data
- *  @param[in] size - Size of the data
- *  @return The checksum
- */
-uint8_t pldm_edac_crc8(const void *data, size_t size);
-
-/** @brief Compute crc32 (same as the one used by IEEE802.3)
- *
- *  @param[in] data - Pointer to the target data
- *  @param[in] size - Size of the data
- *  @return The checksum
- */
-uint32_t pldm_edac_crc32(const void *data, size_t size);
-
-/** @brief Compute cumulative crc32 (same as the one used by IEEE802.3)
- *
- *  @param[in] data - Pointer to the target data
- *  @param[in] size - Size of the data
- *  @param[in] crc - cumulative CRC value
- *  @return The checksum
- */
-uint32_t pldm_edac_crc32_extend(const void *data, size_t size, uint32_t crc);
 
 /** @brief Convert ver32_t to string
  *  @param[in] version - Pointer to ver32_t